use double instead of long to better support unit conversions (without data lost)
git-svn-id: https://svn.apache.org/repos/asf/commons/sandbox/monitoring/branches/modules@731976 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/java/org/apache/commons/monitoring/Counter.java
@@ -18,9 +18,9 @@
 package org.apache.commons.monitoring;
 
 /**
- * A counter to collect application processed items (bytes received, lines processed by a batch,
- * time elapsed by some processing ...).
- *
+ * A counter to collect application processed items (bytes received, lines processed by a batch, time elapsed by some
+ * processing ...).
+ * 
  * @author <a href="mailto:nicolas@apache.org">Nicolas De Loof</a>
  */
 public interface Counter
@@ -29,13 +29,14 @@ public interface Counter
 
     /**
      * Add value to the counter. Delta should not be negative (in such case a Gauge should be used).
+     * 
      * @param delta
      */
-    void add( long delta, Unit unit );
+    void add( double delta, Unit unit );
 
-    public interface Observable extends Metric.Observable<Counter>, Counter
+    public interface Observable
+        extends Metric.Observable<Counter>, Counter
     {
 
     }
 }
-
